Moved these tests under tests/ directory
